+++ CHANGELOG.md
@@ -9,6 +9,15 @@ Whether the application uses the _graphql_, the _generateClientCode_ or the _gen
 * separateUtilityClasses: true _(both client and server mode)_
 * skipGenerationIfSchemaHasNotChanged: true _(both client and server mode)_
 
+
+# 1.18.11
+
+Both modes:
+* Issue #179: Problem with primitive types as java type for custom scalars
+* The build is now Ok with java 17
+* Much better reliability of Subscriptions
+
+
 # 1.18.10
 
 Dependency upgrade: